✨ feat: add sidebar
