+++ src/components/maps/GoogleMap.tsx
@@ -1,10 +1,11 @@
 import { useCallback, useState } from 'react'
 
+import type { GoogleMapProps } from '@react-google-maps/api'
 import { GoogleMap, LoadScript, Marker } from '@react-google-maps/api'
 
 import { cn } from '@/lib/utils'
 
-interface GoogleMapComponentProps {
+interface Props extends GoogleMapProps {
   apiKey: string
   center?: google.maps.LatLngLiteral
   zoom?: number
@@ -14,50 +15,47 @@ interface GoogleMapComponentProps {
   onMapClick?: (event: google.maps.MapMouseEvent) => void
 }
 
-const defaultCenter = {
-  lat: 25.0330,
-  lng: 121.5654, // Taipei, Taiwan
-}
-
-const GoogleMapComponent: React.FC<GoogleMapComponentProps> = ({
+function GoogleMapComponent({
   apiKey,
-  center = defaultCenter,
   zoom = 10,
   markers = [],
   style,
   className,
   onMapClick,
-}) => {
+  ...props
+}: Props) {
   const [map, setMap] = useState<google.maps.Map | null>(null)
 
   const onLoad = useCallback((map: google.maps.Map) => {
     setMap(map)
-    console.log(map)
   }, [])
 
   const onUnmount = useCallback(() => {
+    // 地圖卸載時的清理邏輯
     setMap(null)
   }, [])
 
   return (
-    <LoadScript googleMapsApiKey={apiKey}>
-      <GoogleMap
-        mapContainerStyle={{ ...style }}
-        mapContainerClassName={cn('w-full h-full min-w-full min-h-full', className)}
-        center={center}
-        zoom={zoom}
-        onLoad={onLoad}
-        onUnmount={onUnmount}
-        onClick={onMapClick}
-      >
-        {markers.map((marker, index) => (
-          <Marker
-            key={index}
-            position={marker}
-          />
-        ))}
-      </GoogleMap>
-    </LoadScript>
+    <>
+      <LoadScript googleMapsApiKey={apiKey}>
+        <GoogleMap
+          mapContainerStyle={{ ...style }}
+          mapContainerClassName={cn('size-full', className)}
+          zoom={zoom}
+          onLoad={onLoad}
+          onUnmount={onUnmount}
+          onClick={onMapClick}
+          {...props}
+        >
+          {markers.map(marker => (
+            <Marker
+              key={`${marker.lat}-${marker.lng}`}
+              position={marker}
+            />
+          ))}
+        </GoogleMap>
+      </LoadScript>
+    </>
   )
 }
