+++ src/index.css
@@ -140,6 +140,8 @@
     @apply border-border outline-ring/50;
   }
   body {
-    @apply bg-background text-foreground;
+    @apply text-foreground;
+    background: linear-gradient(135deg, #2a2a72 0%, #009ffd 74%);
   }
 }
+
